chore: unify touch device detection and prune nav deps
